Change generated service to accept version and translate version_date to version in compat layer
