+++ discovery/v1.ts
@@ -27,6 +27,13 @@ class DiscoveryV1 extends GeneratedDiscoveryV1 {
   static VERSION_DATE_2017_04_27: string = '2017-04-27';
   static VERSION_DATE_2016_12_15: string = '2016-12-15';
 
+  constructor(options) {
+    // For backward compatibility, allow version to be passed in version_date.
+    const _options = extend({}, options);
+    _options.version = _options.version_date || _options.version;
+    super(_options);
+  }
+
   static _ensureFilename(file) {
     // no changes needed for streams created by fs.ReadStream (or similar looking streams)
     if (isStream.isReadable(file) && file.path) {
@@ -170,7 +177,7 @@ class DiscoveryV1 extends GeneratedDiscoveryV1 {
   }
 
   query(params, callback) {
-    const _params = params || {};
+    let _params = params || {};
     // query and natural_language_query can't both be populated
     if (_params.query && _params.natural_language_query) {
       delete _params.natural_language_query;
